DECOMPILER: Groovie compiler WIP new files, parsing arguments, overwriting jumps
